C1 to C4 API integration
