+++ app/(tabs)/footer.js
@@ -2,14 +2,14 @@ import React from 'react';
 import { View, Text, StyleSheet,TouchableOpacity } from 'react-native';
 import { LinearGradient } from 'expo-linear-gradient';
 import Icon from 'react-native-vector-icons/Ionicons';
-import { useNavigation } from '@react-navigation/native';
+import { useRouter } from 'expo-router';
 
 export const Footer = () => {
-   const navigation = useNavigation();
+  const router = useRouter();
 
-   const navigateToScreen = (screenName) => {
-     navigation.navigate(screenName);
-   };
+  const navigateToScreen = (screenName) => {
+    router.navigate(screenName); // Use router.push to navigate in Expo Router
+  };
   return (
     <LinearGradient
       start={{ x: 1, y: 0 }} // Start from the right
@@ -19,7 +19,7 @@ export const Footer = () => {
     >
       <TouchableOpacity
         style={styles.iconWrapper}
-        onPress={() => navigateToScreen('ProfileScreen')} // Change 'ProfileScreen' to your actual screen name
+        onPress={() => navigateToScreen('/popularCourses')}// Change 'ProfileScreen' to your actual screen name
       >
         <Icon name="person-outline" size={24} color="white" />
         <Text style={styles.iconLabel}>Profile</Text>
@@ -27,7 +27,7 @@ export const Footer = () => {
 
       <TouchableOpacity
         style={styles.iconWrapper}
-        onPress={() => navigateToScreen('profileCourse')} // Change 'CoursesScreen' to your actual screen name
+        onPress={() => navigateToScreen('/profileHome')} // Change 'CoursesScreen' to your actual screen name
       >
         <Icon name="book-outline" size={24} color="white" />
         <Text style={styles.iconLabel}>Open Courses</Text>
@@ -35,7 +35,7 @@ export const Footer = () => {
 
       <TouchableOpacity
         style={styles.iconWrapper}
-        onPress={() => navigateToScreen('profileHome')} // Change 'AllCoursesScreen' to your actual screen name
+        onPress={() => navigateToScreen('/profileHome')} // Change 'AllCoursesScreen' to your actual screen name
       >
         <Icon name="tablet-portrait-outline" size={24} color="white" />
         <Text style={styles.iconLabel}>Courses</Text>
@@ -43,7 +43,7 @@ export const Footer = () => {
 
       <TouchableOpacity
         style={styles.iconWrapper}
-        onPress={() => navigateToScreen('GamesScreen')} // Change 'GamesScreen' to your actual screen name
+        onPress={() => navigateToScreen('/GamesScreen')} // Change 'GamesScreen' to your actual screen name
       >
         <Icon name="game-controller-outline" size={24} color="white" />
         <Text style={styles.iconLabel}>Games</Text>
@@ -51,7 +51,7 @@ export const Footer = () => {
 
       <TouchableOpacity
         style={styles.iconWrapper}
-        onPress={() => navigateToScreen('SearchScreen')} // Change 'SearchScreen' to your actual screen name
+        onPress={() => navigateToScreen('/SearchScreen')} // Change 'SearchScreen' to your actual screen name
       >
         <Icon name="search-outline" size={24} color="white" />
         <Text style={styles.iconLabel}>Search</Text>
