+++ app/_layout.js
@@ -21,7 +21,7 @@ const RootLayout = () => {
             <Stack.Screen name="(auth)" options={{ headerShown: false }} />
             <Stack.Screen name="index" options={{ headerShown: false }} />
             <Stack.Screen name="(Profile)" options={{ headerShown: false }} />
-
+            <Stack.Screen name="(Course)" options={{ headerShown: false }} />
             {/* <Stack.Screen name="course" component={course} /> */}
             {/* <Stack.Screen name="search/[query]" options={{ headerShown: false }} /> */}
           </Stack>
